prepare the setting changes to allow for multiple command type
